refactor: organize knowledge/ into memory/, transfer/, learning/ subfolders
Move 10 memory-* files to knowledge/memory/.
Move 8 knowledge-transfer-* + transfer-trust files to knowledge/transfer/.
Move 5 learning-* files to knowledge/learning/.
Update all import paths across src/ and tests/.

Also fix adapters/ import paths that were missing from Phase 2.
