+++ src/knowledge/learning/learning-cross-goal.ts
