+++ src/knowledge/memory/memory-query.ts
