+++ src/knowledge/transfer/knowledge-transfer-detect.ts
