+++ tests/e2e/milestone5-semantic.test.ts
@@ -19,8 +19,8 @@ import { StateManager } from "../../src/state-manager.js";
 import { KnowledgeManager } from "../../src/knowledge/knowledge-manager.js";
 import { VectorIndex } from "../../src/knowledge/vector-index.js";
 import { MockEmbeddingClient } from "../../src/knowledge/embedding-client.js";
-import { MemoryLifecycleManager } from "../../src/knowledge/memory-lifecycle.js";
-import type { IDriveScorer } from "../../src/knowledge/memory-lifecycle.js";
+import { MemoryLifecycleManager } from "../../src/knowledge/memory/memory-lifecycle.js";
+import type { IDriveScorer } from "../../src/knowledge/memory/memory-lifecycle.js";
 import { SessionManager } from "../../src/execution/session-manager.js";
 import { GoalDependencyGraph } from "../../src/goal/goal-dependency-graph.js";
 import { ObservationEngine } from "../../src/observation/observation-engine.js";
