+++ tests/knowledge-transfer-apply.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach } from "vitest";
 import * as fs from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
-import { KnowledgeTransfer } from "../src/knowledge/knowledge-transfer.js";
+import { KnowledgeTransfer } from "../src/knowledge/transfer/knowledge-transfer.js";
 import { StateManager } from "../src/state-manager.js";
 import { VectorIndex } from "../src/knowledge/vector-index.js";
 import { MockEmbeddingClient } from "../src/knowledge/embedding-client.js";
