+++ tests/knowledge-transfer-incremental.test.ts
@@ -2,8 +2,8 @@ import { describe, it, expect, vi, beforeEach } from "vitest";
 import * as fs from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
-import { KnowledgeTransfer } from "../src/knowledge/knowledge-transfer.js";
-import { LearningPipeline } from "../src/knowledge/learning-pipeline.js";
+import { KnowledgeTransfer } from "../src/knowledge/transfer/knowledge-transfer.js";
+import { LearningPipeline } from "../src/knowledge/learning/learning-pipeline.js";
 import { ReportingEngine } from "../src/reporting-engine.js";
 import { StateManager } from "../src/state-manager.js";
 import { VectorIndex } from "../src/knowledge/vector-index.js";
