+++ tests/learning-pipeline-feedback.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach, afterEach } from "vitest";
 import * as fs from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
-import { LearningPipeline } from "../src/knowledge/learning-pipeline.js";
+import { LearningPipeline } from "../src/knowledge/learning/learning-pipeline.js";
 import { StateManager } from "../src/state-manager.js";
 import { VectorIndex } from "../src/knowledge/vector-index.js";
 import { MockEmbeddingClient } from "../src/knowledge/embedding-client.js";
