+++ tests/memory-lifecycle-phase2.test.ts
@@ -1,7 +1,7 @@
 import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
 import * as fs from "node:fs";
 import * as path from "node:path";
-import { MemoryLifecycleManager, type IDriveScorer } from "../src/knowledge/memory-lifecycle.js";
+import { MemoryLifecycleManager, type IDriveScorer } from "../src/knowledge/memory/memory-lifecycle.js";
 import type { IEmbeddingClient } from "../src/knowledge/embedding-client.js";
 import type { VectorIndex } from "../src/knowledge/vector-index.js";
 import type { VectorSearchResult } from "../src/types/embedding.js";
