+++ tests/provider-factory.test.ts
@@ -22,15 +22,15 @@ vi.mock("../src/execution/adapter-layer.js", () => ({
   AdapterRegistry: vi.fn().mockImplementation(function() { return { register: vi.fn() }; }),
 }));
 
-vi.mock("../src/adapters/claude-code-cli.js", () => ({
+vi.mock("../src/adapters/agents/claude-code-cli.js", () => ({
   ClaudeCodeCLIAdapter: vi.fn().mockImplementation(function() { return {}; }),
 }));
 
-vi.mock("../src/adapters/claude-api.js", () => ({
+vi.mock("../src/adapters/agents/claude-api.js", () => ({
   ClaudeAPIAdapter: vi.fn().mockImplementation(function() { return {}; }),
 }));
 
-vi.mock("../src/adapters/openai-codex.js", () => ({
+vi.mock("../src/adapters/agents/openai-codex.js", () => ({
   OpenAICodexCLIAdapter: vi.fn().mockImplementation(function() { return {}; }),
 }));
