+++ tests/transfer-trust.test.ts
@@ -2,7 +2,7 @@ import { describe, it, expect, beforeEach } from "vitest";
 import * as fs from "node:fs";
 import * as os from "node:os";
 import * as path from "node:path";
-import { TransferTrustManager } from "../src/knowledge/transfer-trust.js";
+import { TransferTrustManager } from "../src/knowledge/transfer/transfer-trust.js";
 import { StateManager } from "../src/state-manager.js";
 
 // ─── Helpers ───
